Implement vertical alignment.
1. The actual vertical alignment is implemented by adapting the
code of `delarray` to LaTeX3 and to fit the architecture of the
parser.

2. Non-center alignments are implemented by a double pass. This
is the most straightforward way I can think of which guarantees
that we will get correct top/bottom alignment even if the
topmost/bottommost line is embedded in several inner AVMs.

Testing was rudimentary.

Note that bottom alignment is broken if the final row of some AVM
ends in a `\\`.